fix: http 호출 훅 관심사에 맞게 파일 분리
